healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 3841e0fc0474f0d0e7f0a4a85bc346313d5a3dc9
Former-commit-id: 64668f6bb03f546b726b8c95afba23231c9a07a6
Signed-off-by: rajkale99 <kaleraj.rk@gmail.com>